+++ src/main/java/io/github/gaodeliang666/jobtrack/user/context/CurrentUserProvider.java
@@ -1,0 +1,6 @@
+package io.github.gaodeliang666.jobtrack.user.context;
+
+public interface CurrentUserProvider {
+
+    long currentUserId();
+}